removed RungeKutta transcription and all references from the package.  next up is docs
