+++ dymos/__init__.py
@@ -1,7 +1,7 @@
 __version__ = '0.18.1'
 
 from .phase import Phase
-from .transcriptions import GaussLobatto, Radau, RungeKutta
+from .transcriptions import GaussLobatto, Radau
 from .trajectory.trajectory import Trajectory
 from .run_problem import run_problem
 from .load_case import load_case
