+++ dymos/examples/brachistochrone/test/ex_brachistochrone.py
@@ -27,10 +27,6 @@ def brachistochrone_min_time(transcription='gauss-lobatto', num_segments=8, tran
         t = dm.Radau(num_segments=num_segments,
                      order=transcription_order,
                      compressed=compressed)
-    elif transcription == 'runge-kutta':
-        t = dm.RungeKutta(num_segments=num_segments,
-                          order=transcription_order,
-                          compressed=compressed)
     traj = dm.Trajectory()
     phase = dm.Phase(ode_class=BrachistochroneODE, transcription=t)
     p.model.add_subsystem('traj0', traj)
