+++ dymos/examples/brachistochrone/test/test_brachistochrone_control_rate_targets.py
@@ -298,102 +298,6 @@ class TestBrachistochroneControlRateTargets(unittest.TestCase):
 
         plt.show()
 
-    def test_brachistochrone_control_rate_targets_rungekutta(self):
-        import matplotlib.pyplot as plt
-        plt.switch_backend('Agg')
-        import matplotlib.pyplot as plt
-        import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
-        import dymos as dm
-
-        p = om.Problem(model=om.Group())
-        p.driver = om.ScipyOptimizeDriver()
-        p.driver.declare_coloring()
-
-        phase = dm.Phase(ode_class=BrachistochroneRateTargetODE,
-                         transcription=dm.RungeKutta(num_segments=10))
-
-        p.model.add_subsystem('phase0', phase)
-
-        phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
-
-        phase.add_state('x', rate_source='xdot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('y', rate_source='ydot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('v', rate_source='vdot',
-                        units='m/s',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_control('theta', units='deg*s', lower=0.01, upper=179.9, fix_initial=True)
-
-        phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
-
-        phase.add_boundary_constraint('x', loc='final', equals=10)
-        phase.add_boundary_constraint('y', loc='final', equals=5)
-
-        # Minimize time at the end of the phase
-        phase.add_objective('time', loc='final', scaler=10)
-
-        p.model.linear_solver = om.DirectSolver()
-
-        p.setup()
-
-        p['phase0.t_initial'] = 0.0
-        p['phase0.t_duration'] = 2.0
-
-        p['phase0.states:x'] = phase.interpolate(ys=[0, 10], nodes='state_input')
-        p['phase0.states:y'] = phase.interpolate(ys=[10, 5], nodes='state_input')
-        p['phase0.states:v'] = phase.interpolate(ys=[0, 9.9], nodes='state_input')
-        p['phase0.controls:theta'] = phase.interpolate(ys=[0, 100], nodes='control_input')
-
-        # Solve for the optimal trajectory
-        p.run_driver()
-
-        # Test the results
-        assert_near_equal(p.get_val('phase0.timeseries.time')[-1], 1.8016, tolerance=1.0E-3)
-
-        # Generate the explicitly simulated trajectory
-        exp_out = phase.simulate()
-
-        fig, ax = plt.subplots()
-        fig.suptitle('Brachistochrone Solution')
-
-        x_imp = p.get_val('phase0.timeseries.states:x')
-        y_imp = p.get_val('phase0.timeseries.states:y')
-
-        x_exp = exp_out.get_val('phase0.timeseries.states:x')
-        y_exp = exp_out.get_val('phase0.timeseries.states:y')
-
-        ax.plot(x_imp, y_imp, 'ro', label='solution')
-        ax.plot(x_exp, y_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('x (m)')
-        ax.set_ylabel('y (m)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        fig, ax = plt.subplots()
-
-        t_imp = p.get_val('phase0.timeseries.time')
-        theta_imp = p.get_val('phase0.timeseries.control_rates:theta_rate')
-        t_exp = exp_out.get_val('phase0.timeseries.time')
-        theta_exp = exp_out.get_val('phase0.timeseries.control_rates:theta_rate')
-
-        ax.plot(t_imp, theta_imp, 'ro', label='solution')
-        ax.plot(t_exp, theta_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('time (s)')
-        ax.set_ylabel(r'$\theta$ (deg)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        plt.show()
-
 
 @use_tempdirs
 class TestBrachistochroneExplicitControlRateTargets(unittest.TestCase):
@@ -526,103 +430,6 @@ class TestBrachistochroneExplicitControlRateTargets(unittest.TestCase):
                           rate_targets=['theta_rate'])
 
         phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
-
-        # Minimize time at the end of the phase
-        phase.add_objective('time', loc='final', scaler=10)
-
-        p.model.linear_solver = om.DirectSolver()
-
-        p.setup()
-
-        p['phase0.t_initial'] = 0.0
-        p['phase0.t_duration'] = 2.0
-
-        p['phase0.states:x'] = phase.interpolate(ys=[0, 10], nodes='state_input')
-        p['phase0.states:y'] = phase.interpolate(ys=[10, 5], nodes='state_input')
-        p['phase0.states:v'] = phase.interpolate(ys=[0, 9.9], nodes='state_input')
-        p['phase0.controls:theta'] = phase.interpolate(ys=[0, 100], nodes='control_input')
-
-        # Solve for the optimal trajectory
-        p.run_driver()
-
-        # Test the results
-        assert_near_equal(p.get_val('phase0.timeseries.time')[-1], 1.8016, tolerance=1.0E-3)
-
-        # Generate the explicitly simulated trajectory
-        exp_out = phase.simulate()
-
-        fig, ax = plt.subplots()
-        fig.suptitle('Brachistochrone Solution')
-
-        x_imp = p.get_val('phase0.timeseries.states:x')
-        y_imp = p.get_val('phase0.timeseries.states:y')
-
-        x_exp = exp_out.get_val('phase0.timeseries.states:x')
-        y_exp = exp_out.get_val('phase0.timeseries.states:y')
-
-        ax.plot(x_imp, y_imp, 'ro', label='solution')
-        ax.plot(x_exp, y_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('x (m)')
-        ax.set_ylabel('y (m)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        fig, ax = plt.subplots()
-
-        t_imp = p.get_val('phase0.timeseries.time')
-        theta_imp = p.get_val('phase0.timeseries.control_rates:theta_rate')
-        t_exp = exp_out.get_val('phase0.timeseries.time')
-        theta_exp = exp_out.get_val('phase0.timeseries.control_rates:theta_rate')
-
-        ax.plot(t_imp, theta_imp, 'ro', label='solution')
-        ax.plot(t_exp, theta_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('time (s)')
-        ax.set_ylabel(r'$\theta$ (deg)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        plt.show()
-
-    def test_brachistochrone_control_rate_targets_rungekutta(self):
-        import matplotlib.pyplot as plt
-        plt.switch_backend('Agg')
-        import matplotlib.pyplot as plt
-        import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
-        import dymos as dm
-
-        p = om.Problem(model=om.Group())
-        p.driver = om.ScipyOptimizeDriver()
-        p.driver.declare_coloring()
-
-        phase = dm.Phase(ode_class=BrachistochroneRateTargetODE,
-                         transcription=dm.RungeKutta(num_segments=10))
-
-        p.model.add_subsystem('phase0', phase)
-
-        phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
-
-        phase.add_state('x', rate_source='xdot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('y', rate_source='ydot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('v', rate_source='vdot',
-                        units='m/s',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_control('theta', units='deg*s', lower=0.01, upper=179.9, fix_initial=True,
-                          rate_targets=['theta_rate'])
-
-        phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
-
-        phase.add_boundary_constraint('x', loc='final', equals=10)
-        phase.add_boundary_constraint('y', loc='final', equals=5)
 
         # Minimize time at the end of the phase
         phase.add_objective('time', loc='final', scaler=10)
@@ -873,103 +680,6 @@ class TestBrachistochronePolynomialControlRateTargets(unittest.TestCase):
 
         plt.show()
 
-    def test_brachistochrone_polynomial_control_rate_targets_rungekutta(self):
-        import matplotlib.pyplot as plt
-        plt.switch_backend('Agg')
-        import matplotlib.pyplot as plt
-        import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
-        import dymos as dm
-
-        p = om.Problem(model=om.Group())
-        p.driver = om.ScipyOptimizeDriver()
-        p.driver.declare_coloring()
-
-        phase = dm.Phase(ode_class=BrachistochroneRateTargetODE,
-                         transcription=dm.RungeKutta(num_segments=10))
-
-        p.model.add_subsystem('phase0', phase)
-
-        phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
-
-        phase.add_state('x', rate_source='xdot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('y', rate_source='ydot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('v', rate_source='vdot',
-                        units='m/s',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_polynomial_control('theta', order=3, units='deg*s', lower=0.01, upper=179.9,
-                                     fix_initial=True, val=1.0)
-
-        phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
-
-        phase.add_boundary_constraint('x', loc='final', equals=10)
-        phase.add_boundary_constraint('y', loc='final', equals=5)
-
-        # Minimize time at the end of the phase
-        phase.add_objective('time', loc='final', scaler=10)
-
-        p.model.linear_solver = om.DirectSolver()
-
-        p.setup()
-
-        p['phase0.t_initial'] = 0.0
-        p['phase0.t_duration'] = 2.0
-
-        p['phase0.states:x'] = phase.interpolate(ys=[0, 10], nodes='state_input')
-        p['phase0.states:y'] = phase.interpolate(ys=[10, 5], nodes='state_input')
-        p['phase0.states:v'] = phase.interpolate(ys=[0, 9.9], nodes='state_input')
-        p['phase0.polynomial_controls:theta'] = [0, 10, 80, 100]
-
-        # Solve for the optimal trajectory
-        p.run_driver()
-
-        # Test the results
-        assert_near_equal(p.get_val('phase0.timeseries.time')[-1], 1.8016, tolerance=1.0E-3)
-
-        # Generate the explicitly simulated trajectory
-        exp_out = phase.simulate()
-
-        fig, ax = plt.subplots()
-        fig.suptitle('Brachistochrone Solution')
-
-        x_imp = p.get_val('phase0.timeseries.states:x')
-        y_imp = p.get_val('phase0.timeseries.states:y')
-
-        x_exp = exp_out.get_val('phase0.timeseries.states:x')
-        y_exp = exp_out.get_val('phase0.timeseries.states:y')
-
-        ax.plot(x_imp, y_imp, 'ro', label='solution')
-        ax.plot(x_exp, y_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('x (m)')
-        ax.set_ylabel('y (m)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        fig, ax = plt.subplots()
-
-        t_imp = p.get_val('phase0.timeseries.time')
-        theta_imp = p.get_val('phase0.timeseries.polynomial_control_rates:theta_rate')
-        t_exp = exp_out.get_val('phase0.timeseries.time')
-        theta_exp = exp_out.get_val('phase0.timeseries.polynomial_control_rates:theta_rate')
-
-        ax.plot(t_imp, theta_imp, 'ro', label='solution')
-        ax.plot(t_exp, theta_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('time (s)')
-        ax.set_ylabel(r'$\theta$ (deg)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        plt.show()
-
 
 @use_tempdirs
 class TestBrachistochronePolynomialControlExplicitRateTargets(unittest.TestCase):
@@ -1161,103 +871,6 @@ class TestBrachistochronePolynomialControlExplicitRateTargets(unittest.TestCase)
 
         plt.show()
 
-    def test_brachistochrone_polynomial_control_rate_targets_rungekutta(self):
-        import matplotlib.pyplot as plt
-        plt.switch_backend('Agg')
-        import matplotlib.pyplot as plt
-        import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
-        import dymos as dm
-
-        p = om.Problem(model=om.Group())
-        p.driver = om.ScipyOptimizeDriver()
-        p.driver.declare_coloring()
-
-        phase = dm.Phase(ode_class=BrachistochroneRateTargetODE,
-                         transcription=dm.RungeKutta(num_segments=10))
-
-        p.model.add_subsystem('phase0', phase)
-
-        phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
-
-        phase.add_state('x', rate_source='xdot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('y', rate_source='ydot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('v', rate_source='vdot',
-                        units='m/s',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_polynomial_control('theta', order=3, units='deg*s', lower=0.01, upper=179.9,
-                                     rate_targets=['theta_rate'], fix_initial=True)
-
-        phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
-
-        phase.add_boundary_constraint('x', loc='final', equals=10)
-        phase.add_boundary_constraint('y', loc='final', equals=5)
-
-        # Minimize time at the end of the phase
-        phase.add_objective('time', loc='final', scaler=10)
-
-        p.model.linear_solver = om.DirectSolver()
-
-        p.setup()
-
-        p['phase0.t_initial'] = 0.0
-        p['phase0.t_duration'] = 2.0
-
-        p['phase0.states:x'] = phase.interpolate(ys=[0, 10], nodes='state_input')
-        p['phase0.states:y'] = phase.interpolate(ys=[10, 5], nodes='state_input')
-        p['phase0.states:v'] = phase.interpolate(ys=[0, 9.9], nodes='state_input')
-        p['phase0.polynomial_controls:theta'] = [0, 10, 80, 100]
-
-        # Solve for the optimal trajectory
-        p.run_driver()
-
-        # Test the results
-        assert_near_equal(p.get_val('phase0.timeseries.time')[-1], 1.8016, tolerance=1.0E-3)
-
-        # Generate the explicitly simulated trajectory
-        exp_out = phase.simulate()
-
-        fig, ax = plt.subplots()
-        fig.suptitle('Brachistochrone Solution')
-
-        x_imp = p.get_val('phase0.timeseries.states:x')
-        y_imp = p.get_val('phase0.timeseries.states:y')
-
-        x_exp = exp_out.get_val('phase0.timeseries.states:x')
-        y_exp = exp_out.get_val('phase0.timeseries.states:y')
-
-        ax.plot(x_imp, y_imp, 'ro', label='solution')
-        ax.plot(x_exp, y_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('x (m)')
-        ax.set_ylabel('y (m)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        fig, ax = plt.subplots()
-
-        t_imp = p.get_val('phase0.timeseries.time')
-        theta_imp = p.get_val('phase0.timeseries.polynomial_control_rates:theta_rate')
-        t_exp = exp_out.get_val('phase0.timeseries.time')
-        theta_exp = exp_out.get_val('phase0.timeseries.polynomial_control_rates:theta_rate')
-
-        ax.plot(t_imp, theta_imp, 'ro', label='solution')
-        ax.plot(t_exp, theta_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('time (s)')
-        ax.set_ylabel(r'$\theta$ (deg)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        plt.show()
-
 
 @use_tempdirs
 class TestBrachistochronePolynomialControlExplicitRate2Targets(unittest.TestCase):
@@ -1296,103 +909,6 @@ class TestBrachistochronePolynomialControlExplicitRate2Targets(unittest.TestCase
                                      rate_targets=None, rate2_targets=['theta_rate'], fix_initial=True)
 
         phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
-
-        # Minimize time at the end of the phase
-        phase.add_objective('time', loc='final', scaler=10)
-
-        p.model.linear_solver = om.DirectSolver()
-
-        p.setup()
-
-        p['phase0.t_initial'] = 0.0
-        p['phase0.t_duration'] = 2.0
-
-        p['phase0.states:x'] = phase.interpolate(ys=[0, 10], nodes='state_input')
-        p['phase0.states:y'] = phase.interpolate(ys=[10, 5], nodes='state_input')
-        p['phase0.states:v'] = phase.interpolate(ys=[0, 9.9], nodes='state_input')
-        p['phase0.polynomial_controls:theta'] = [0, 10, 40, 60, 80, 100]
-
-        # Solve for the optimal trajectory
-        p.run_driver()
-
-        # Test the results
-        assert_near_equal(p.get_val('phase0.timeseries.time')[-1], 1.8016, tolerance=1.0E-3)
-
-        # Generate the explicitly simulated trajectory
-        exp_out = phase.simulate()
-
-        fig, ax = plt.subplots()
-        fig.suptitle('Brachistochrone Solution')
-
-        x_imp = p.get_val('phase0.timeseries.states:x')
-        y_imp = p.get_val('phase0.timeseries.states:y')
-
-        x_exp = exp_out.get_val('phase0.timeseries.states:x')
-        y_exp = exp_out.get_val('phase0.timeseries.states:y')
-
-        ax.plot(x_imp, y_imp, 'ro', label='solution')
-        ax.plot(x_exp, y_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('x (m)')
-        ax.set_ylabel('y (m)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        fig, ax = plt.subplots()
-
-        t_imp = p.get_val('phase0.timeseries.time')
-        theta_imp = p.get_val('phase0.timeseries.polynomial_control_rates:theta_rate')
-        t_exp = exp_out.get_val('phase0.timeseries.time')
-        theta_exp = exp_out.get_val('phase0.timeseries.polynomial_control_rates:theta_rate')
-
-        ax.plot(t_imp, theta_imp, 'ro', label='solution')
-        ax.plot(t_exp, theta_exp, 'b-', label='simulated')
-
-        ax.set_xlabel('time (s)')
-        ax.set_ylabel(r'$\theta$ (deg)')
-        ax.grid(True)
-        ax.legend(loc='upper right')
-
-        plt.show()
-
-    def test_brachistochrone_polynomial_control_rate_targets_rungekutta(self):
-        import matplotlib.pyplot as plt
-        plt.switch_backend('Agg')
-        import matplotlib.pyplot as plt
-        import openmdao.api as om
-        from openmdao.utils.assert_utils import assert_near_equal
-        import dymos as dm
-
-        p = om.Problem(model=om.Group())
-        p.driver = om.ScipyOptimizeDriver()
-        p.driver.declare_coloring()
-
-        phase = dm.Phase(ode_class=BrachistochroneRateTargetODE,
-                         transcription=dm.RungeKutta(num_segments=10))
-
-        p.model.add_subsystem('phase0', phase)
-
-        phase.set_time_options(initial_bounds=(0, 0), duration_bounds=(.5, 10))
-
-        phase.add_state('x', rate_source='xdot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('y', rate_source='ydot',
-                        units='m',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_state('v', rate_source='vdot',
-                        units='m/s',
-                        fix_initial=True, fix_final=False, solve_segments=False)
-
-        phase.add_polynomial_control('theta', order=5, units='deg*s**2', lower=0.01, upper=179.9,
-                                     rate_targets=None, rate2_targets=['theta_rate'], fix_initial=True)
-
-        phase.add_parameter('g', units='m/s**2', opt=False, val=9.80665)
-
-        phase.add_boundary_constraint('x', loc='final', equals=10)
-        phase.add_boundary_constraint('y', loc='final', equals=5)
 
         # Minimize time at the end of the phase
         phase.add_objective('time', loc='final', scaler=10)
