+++ dymos/examples/brachistochrone/test/test_brachistochrone_solve_segments.py
@@ -36,10 +36,7 @@ def _make_problem(transcription='gauss-lobatto', num_segments=8, transcription_o
         t = dm.Radau(num_segments=num_segments,
                      order=transcription_order,
                      compressed=compressed)
-    elif transcription == 'runge-kutta':
-        t = dm.RungeKutta(num_segments=num_segments,
-                          order=transcription_order,
-                          compressed=compressed)
+
     traj = dm.Trajectory()
     phase = dm.Phase(ode_class=BrachistochroneODE, transcription=t)
     p.model.add_subsystem('traj0', traj)
