+++ dymos/examples/brachistochrone/test/test_ex_brachistochrone_vector_states.py
@@ -86,15 +86,6 @@ class TestBrachistochroneVectorStatesExample(unittest.TestCase):
         self.assert_results(p)
         self.assert_partials(p)
 
-    def test_ex_brachistochrone_vs_rungekutta_compressed(self):
-        p = ex_brachistochrone_vs.brachistochrone_min_time(transcription='runge-kutta',
-                                                           transcription_order='RK4',
-                                                           compressed=True,
-                                                           force_alloc_complex=True,
-                                                           run_driver=True)
-
-        self.assert_results(p)
-
 
 if __name__ == "__main__":
     unittest.main()
