+++ dymos/examples/vanderpol/vanderpol_dymos.py
@@ -35,10 +35,6 @@ def vanderpol(transcription='gauss-lobatto', num_segments=8, transcription_order
         t = dm.Radau(num_segments=num_segments,
                      order=transcription_order,
                      compressed=compressed)
-    elif transcription == 'runge-kutta':
-        t = dm.RungeKutta(num_segments=num_segments,
-                          order=transcription_order,
-                          compressed=compressed)
 
     # define a Phase as specified above and add to Phase
     if not delay:
