+++ dymos/test/test_options.py
@@ -23,14 +23,6 @@ class TestOptions(unittest.TestCase):
         self.assertSetEqual(set(cpd.keys()), {'traj0.phases.phase0.rhs_disc',
                                               'traj0.phases.phase0.rhs_col'})
 
-    def test_include_check_partials_false_rk(self):
-        dm.options['include_check_partials'] = False
-        p = brachistochrone_min_time(transcription='runge-kutta', compressed=False,
-                                     run_driver=False, force_alloc_complex=True)
-        cpd = p.check_partials(out_stream=None, method='fd')
-        self.assertSetEqual(set(cpd.keys()), {'traj0.phases.phase0.ode',
-                                              'traj0.phases.phase0.rk_solve_group.k_iter_group.ode'})
-
     def test_include_check_partials_true_radau(self):
         dm.options['include_check_partials'] = True
         p = brachistochrone_min_time(transcription='radau-ps', compressed=False,
@@ -41,13 +33,6 @@ class TestOptions(unittest.TestCase):
     def test_include_check_partials_true_gl(self):
         dm.options['include_check_partials'] = True
         p = brachistochrone_min_time(transcription='gauss-lobatto', compressed=False,
-                                     run_driver=False, force_alloc_complex=True)
-        cpd = p.check_partials(out_stream=None, method='fd')
-        self.assertTrue(len(list(cpd.keys())) > 1)
-
-    def test_include_check_partials_true_rk(self):
-        dm.options['include_check_partials'] = True
-        p = brachistochrone_min_time(transcription='runge-kutta', compressed=False,
                                      run_driver=False, force_alloc_complex=True)
         cpd = p.check_partials(out_stream=None, method='fd')
         self.assertTrue(len(list(cpd.keys())) > 1)
