+++ dymos/transcriptions/__init__.py
@@ -1,4 +1,3 @@
-from .runge_kutta.runge_kutta import RungeKutta
 from .pseudospectral.gauss_lobatto import GaussLobatto
 from .pseudospectral.radau_pseudospectral import Radau
 from .solve_ivp.solve_ivp import SolveIVP
