+++ dymos/transcriptions/pseudospectral/components/test/test_collocation_defect_solver.py
@@ -305,10 +305,7 @@ class TestCollocationBalanceApplyNL(unittest.TestCase):
             t = dm.Radau(num_segments=num_segments,
                          order=transcription_order,
                          compressed=compressed)
-        elif transcription == 'runge-kutta':
-            t = dm.RungeKutta(num_segments=num_segments,
-                              order=transcription_order,
-                              compressed=compressed)
+
         traj = dm.Trajectory()
         phase = dm.Phase(ode_class=BrachistochroneODE, transcription=t)
         p.model.add_subsystem('traj0', traj)
