+++ dymos/transcriptions/solve_ivp/components/test/test_segment_simulation_comp.py
@@ -1,17 +1,40 @@
 import unittest
 
+import numpy as np
 import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 from openmdao.utils.testing_utils import use_tempdirs
 
 from dymos.transcriptions.solve_ivp.components.segment_simulation_comp import SegmentSimulationComp
-from dymos.transcriptions.runge_kutta.test.rk_test_ode import TestODE
 from dymos.phase.options import TimeOptionsDictionary, StateOptionsDictionary
 from dymos.transcriptions.grid_data import GridData
 
 # Modify class so we can run it standalone.
 from dymos.utils.misc import CompWrapperConfig
 SegmentSimulationComp = CompWrapperConfig(SegmentSimulationComp)
+
+
+class TestODE(om.ExplicitComponent):
+
+    def initialize(self):
+        self.options.declare('num_nodes', types=int)
+
+    def setup(self):
+        self.add_input('t', val=np.ones(self.options['num_nodes']), units='s')
+        self.add_input('y', val=np.ones(self.options['num_nodes']), units='m')
+        self.add_output('ydot', val=np.ones(self.options['num_nodes']), units='m/s')
+
+        ar = np.arange(self.options['num_nodes'])
+        self.declare_partials(of='ydot', wrt='t', rows=ar, cols=ar)
+        self.declare_partials(of='ydot', wrt='y', rows=ar, cols=ar, val=1.0)
+
+    def compute(self, inputs, outputs, discrete_inputs=None, discrete_outputs=None):
+        t = inputs['t']
+        y = inputs['y']
+        outputs['ydot'] = y - t ** 2 + 1
+
+    def compute_partials(self, inputs, partials):
+        partials['ydot', 't'] = -2 * inputs['t']
 
 
 @use_tempdirs
